requirements.txt
# Chainlytics Backend ML System - Requirements

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0
lightgbm>=3.2.0
torch>=1.9.0
torchvision>=0.10.0

# Deep learning (optional for advanced models)
tensorflow>=2.6.0
tensorboard>=2.6.0

# Graph neural networks (for state encoding)
torch-geometric>=2.0.0

# Optimization
ortools>=9.0.0

# Data processing and validation
pyyaml>=5.4.0
jsonschema>=3.2.0

# Explainability
shap>=0.39.0

# Database connectivity
psycopg2-binary>=2.9.0
redis>=3.5.0

# Logging and monitoring
loguru>=0.5.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Utilities
tqdm>=4.62.0
joblib>=1.0.0

# Async support
asyncio-mqtt>=0.10.0

# Development tools
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# Visualization (for analysis)
matplotlib>=3.4.0
seaborn>=0.11.0

# API (if needed for inference endpoints)
fastapi>=0.68.0
uvicorn>=0.15.0

# Containerization support
docker>=5.0.0